==> notebook/2_rag_for_drug_repurposing.ipynb <==
### Unstructured Data

!head -100 ../data/unstructured/unstructured_gene_disease.txt

from llama_index import Document
from llama_index import SimpleDirectoryReader

documents = SimpleDirectoryReader(
    input_files=["../data/unstructured/unstructured_gene_disease.txt"]
).load_data()

document = Document(text="\n\n".join([doc for doc in documents[0].text.split('\n')]))

print(f'document type: {type(documents)}')
print(f'total # docs: {len(documents)}')
print(f'doc type: {type(documents[0])}')
print(f'doc sample: {documents[0]}')

import torch

from llama_index import ServiceContext
from llama_index import VectorStoreIndex
from llama_index.llms import HuggingFaceLLM
from llama_index.prompts import PromptTemplate

# This will wrap the default prompts that are internal to llama-index
# taken from https://huggingface.co/Writer/camel-5b-hf
query_wrapper_prompt = PromptTemplate(
    "Below is an instruction that describes a task. "
    "Write a response that appropriately completes the request.\n\n"
    "### Instruction:\n{query_str}\n\n### Response:"
)

llm = HuggingFaceLLM(
    context_window=2048,
    max_new_tokens=256,
    generate_kwargs={"temperature": 0.25, "do_sample": False},
    query_wrapper_prompt=query_wrapper_prompt,
    tokenizer_name="Writer/camel-5b-hf",
    model_name="Writer/camel-5b-hf",
    device_map="auto",
    tokenizer_kwargs={"max_length": 2048},
    # model_kwargs={"torch_dtype": torch.float16}
)

service_context = ServiceContext.from_defaults(
    chunk_size=512,
    llm=llm,
    embed_model="local:BAAI/bge-small-en-v1.5"
)
index = VectorStoreIndex.from_documents([document],
                                        service_context=service_context)
query_engine = index.as_query_engine()

response = query_engine.query(
    "Describe the role of Gene::10 in different diseases."
)
print(str(response))

response = query_engine.query(
    "What is the top gene that causes disease? Describe the relationships with the disease."
)
print(str(response))

response = query_engine.query(
    "List top 5 genes that is associated with Disease::MESH:D014947? Briefly describe how they are associated."
)
print(str(response))

### Semi-structured Data

!head -100 ../data/semi_structured/semi_structured_gene_disease.txt

documents = SimpleDirectoryReader(
    input_files=["../data/semi_structured/semi_structured_gene_disease.txt"]
).load_data()

document = Document(text="\n\n".join([doc for doc in documents[0].text.split('\n')]))

print(f'document type: {type(documents)}')
print(f'total # docs: {len(documents)}')
print(f'doc type: {type(documents[0])}')
print(f'doc sample: {documents[0]}')

index = VectorStoreIndex.from_documents([document],
                                        service_context=service_context)
query_engine = index.as_query_engine()

response = query_engine.query(
    "Describe the role of Gene::10 in different diseases."
)
print(str(response))

response = query_engine.query(
    "What is the top gene that causes disease? Describe the relationships with the disease."
)
print(str(response))

response = query_engine.query(
    "List top 5 genes that is associated with Disease::MESH:D014947? Briefly describe how they are associated."
)
print(str(response))

### Structured Data

!head -100 ../data/structured/structured_gene_disease.csv

documents = SimpleDirectoryReader(
    input_files=["../data/structured/structured_gene_disease.csv"]
).load_data()

document = Document(text="\n\n".join([doc for doc in documents[0].text.split('\n')]))

print(f'document type: {type(documents)}')
print(f'total # docs: {len(documents)}')
print(f'doc type: {type(documents[0])}')
print(f'doc sample: {documents[0]}')

index = VectorStoreIndex.from_documents([document],
                                        service_context=service_context)
query_engine = index.as_query_engine()

response = query_engine.query(
    "Describe the role of Gene::10 in different diseases."
)
print(str(response))

response = query_engine.query(
    "What is the top gene that causes disease? Describe the relationships with the disease."
)
print(str(response))

response = query_engine.query(
    "List top 5 genes that is associated with Disease::MESH:D014947? Briefly describe how they are associated."
)
print(str(response))
