==> notebook/3_create_embeddings (1).ipynb <==
import pandas as pd
import pickle

!pip install gensim
#gensim needs to be installed in current environment

from gensim.models import Word2Vec
#no need to tokenize our data cause there are no sentences, just a bunch of illogical nouns




#we create embeddings on the previously created dataset, data.csv
#use word2vec?
#pick appropriate embedding model not whatever we think is popular
#https://medium.com/@mehdi_chebbah/how-to-choose-the-right-embedding-model-for-you-1fc917d14517 -> based on this article, we have to pick static word
#level embedding model->word2vec or GloVe
#using word2vec for now

df = pd.read_csv('../data/structured/groupby_data.csv')
df.head()


df.e2 = df.e2.apply(eval)
#e2 is being read as strings, just making them lists

# model1 = Word2Vec([["akhil","akhi"]*70,["surya","parryhotter"]*90,["goparaju","naga","GN","genie"]*50], vector_size=30, window=5, min_count=0, workers=5,epochs=10,sorted_vocab=0)
# model1.wv.index_to_key


# model1.wv.most_similar('genie',topn=5)
#just checking syntax and output formats

model = Word2Vec(df.e2, vector_size=50, window=10, min_count=1, workers=5,epochs=10)
model.wv.most_similar('db00014',topn=5)



for j in df.head(5).e1:
    print(j)
    for k,i in enumerate(model.wv.most_similar(j,topn=100)):
        if i[0] in df.iloc[0,1]:
            print(k,i)

#the results are not promising, are embeddings supposed to work for this?


df.e2[df.e2.apply(lambda x: 'db00644' in x)]
# well we cannot certainly say as the same compound may have appeared multiple times and may be the most similar compound from the above has been associated
#more often with current compound

df.e2[df.e2.apply(lambda x: 'db00014' in x and 'db00666' in x)]

df.e2[df.e2.apply(lambda x: 'db00014' in x and 'db00644' in x)]

#the above two rows sort of support what i thought, the data suggests those two compounds are associated.... this is purely on the basis of
#the number of occurances, i might be gravely wrong... however whoever made the initial knowledge graph dataset might have analysed these compounds at
#molecular level and concluded that they are similar in some way

#i think i should create a dictionary with the side effect as key and known compounds as values...
#our question would be  "what untested compound could have caused this side effect". -> we would get the top 10 similar compounds of each value and then select whichever 
#is the highest and does not occur in the value

side_eff_dict = {}
se = pd.read_csv('../data/structured/side_effects.csv')
se.head(),se.shape

for i,j in zip(se.side_effect,se.compound):
    if j in side_eff_dict.keys():
        side_eff_dict[j].append(i)
    else:
        side_eff_dict[j]=[i]


len(side_eff_dict)

#len(se.e2.value_counts()) #got 5701

for i in side_eff_dict.keys():
    side_eff_dict[i]=set(side_eff_dict[i])
    #to have faster indexing

with open('../data/meta/side_effects.pickle', 'wb') as file:
    pickle.dump(side_eff_dict, file)

model.save('../data/meta/compound_embeddings.bin')
