==> notebook/4_find_possible_compound.ipynb <==
import pandas as pd
import pickle
from math import e

from gensim.models import Word2Vec


# in his notebook we will find possible compounds which could cause a given side effect.\
#using the sigmoid function to represent the similarity of the untested compound and known compound causing the side effect


with open('../data/meta/side_effects.pickle', 'rb') as file:
    side_eff_dict = pickle.load(file)


df=pd.DataFrame(side_eff_dict.items(),columns=['side_effect','compounds'])
df.head()

model=Word2Vec.load('../data/meta/compound_embeddings.bin')

model.wv.most_similar('db00014',topn=5)

def get_similar_compounds(compound,top=5):
    return model.wv.most_similar(compound,topn=top)
def sigm(a):
    return (1+e**-a)**-1
def wrapper(col2,top=5):
    lis =[]
   
    for comp in col2:
        lis += get_similar_compounds(comp)
        #insertion takes n and sorting is n logn, the current values of n are small enough to make the cost negligible.
        #we will just sort at the end of the function and return only the top n compounds.
    top_lis = []
    keys={}
    for i in lis:
        if (not i[0] in col2) and (not i[0] in keys.keys()):
            top_lis.append(i)
        keys[i[0]] = 1
    
    top_lis.sort(key=lambda x:x[1],reverse=True)
    return top_lis[0][0],sigm(top_lis[0][1]),[i[0] for i in top_lis[1:top+1]]

            
        


df[['mostProbableUntestedCompound','score','otherPossibleCompounds']] = df.compounds.apply(lambda x:pd.Series(wrapper(x)))
df.head()

df.to_csv('../data/results/side_effect_compound_pred.csv',index=False)

#score is sigmoid of the similarity (which is a cosine similarity)

#we initially thought of using graph rags and a graph database and then create texts based on those.
#however there are no graph databases which offer such services without some kind of subscription. So I changed my approach and thought of using embeddings.
#i didnt use LLM because the model will focus on the text patterns and grammar more than on the compound names. However i want to see if what i 
#think is true by using the story based data with a very structured format, all rows would have only have 2 patterns which might make 
#it hard for the model to focus on the compound ids.
