==> notebook/5_LLM.ipynb <==
import pandas as pd

df = pd.read_csv('../data/unstructured/llm_data.csv')

from transformers import AutoModelForMaskedLM
from transformers import AutoTokenizer
from transformers import DataCollatorForLanguageModeling
# from transformers import Trainer, TrainingArguments

#ref-> https://colab.research.google.com/github/huggingface/notebooks/blob/master/examples/language_modeling.ipynb


# !pip uninstall bitsandbytes -y

# !pip uninstall bitsandbytes -y

# !pip install bitsandbytes==0.41.0

# !pip freeze

from transformers import Trainer, TrainingArguments


model_name = "distilroberta-base"
model = AutoModelForMaskedLM.from_pretrained(model_name)
tokenizer = AutoTokenizer.from_pretrained(model_name, use_fast=True)

tokenized_dataset = df.text.apply(tokenizer)


data_collator = DataCollatorForLanguageModeling(tokenizer=tokenizer, mlm_probability=0.15)


training_args = TrainingArguments(
    output_dir='..data/meta/distil_roberta_chkpts/',
    # evaluation_strategy = "epoch",
    learning_rate=2e-5,
    weight_decay=0.01,
    
)

trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=tokenized_dataset,
    
    data_collator=data_collator,
)

trainer.train()
#takes around 40 minutes with one gpu

model.save_pretrained('../data/meta/model-distilrobertabase/')

m1 = AutoModelForMaskedLM.from_pretrained('../data/meta/model-distilrobertabase/')
t1 =  AutoTokenizer.from_pretrained("distilroberta-base")
#we are using the tokenizer from distilroberta-base

from transformers import pipeline
unmasker = pipeline('fill-mask', model='../data/meta/model-distilrobertabase/',tokenizer = 'distilroberta-base')
unmasker("db00014 is similar to db<mask>.")



# we can see that the tokenizer is tokenizing the compound name which is undesirable and then it is
#focusing on the english words 'to' and 'similar' more than the compound name.

unmasker("side effect c0038990 caused by db<mask>")

# we can see that the model is giving random outputs which are not relevant and useful.
# i think i should have trained it from scratch instead of finetuning as this is a very domain specefic task and the models' 
# previous knowledge is not exactly useful.
# i could have used a clinical model for this task but it might take a lot of time to pre-train and use a larger model.

unmasker("db00014 is similar to db<mask>.")

# we can see the tokenizer is at the blame again, it is removing the zeros to the left which is not desirable.we should not split 
# the ids and should wor at word level
# we have to make a custom tokneizer to prevent this.

# I did not find LLMs to be useful for this task because of the nature of the dataset, its not textual and does not include notes or comments
# to help the model understand from a bigger pool of data and to provide a need to use LLM instead of machine learning models.
